:hammer: Move shared utils implementation to commonMain
