--- composeApp/src/desktopMain/kotlin/com/crosspaste/utils/ControlUtils.desktop.kt
@@ -1,78 +1,0 @@
-package com.crosspaste.utils
-
-import kotlinx.coroutines.delay
-
-actual fun getControlUtils(): ControlUtils {
-    return DesktopControlUtils
-}
-
-object DesktopControlUtils : ControlUtils {
-
-    override suspend fun <T> ensureMinExecutionTime(
-        delayTime: Int,
-        action: suspend () -> T,
-    ): T {
-        val start = System.currentTimeMillis()
-        val result = action()
-        val end = System.currentTimeMillis()
-
-        val remainingDelay = delayTime + start - end
-
-        if (remainingDelay > 0) {
-            delay(remainingDelay)
-        }
-        return result
-    }
-
-    override suspend fun <T> exponentialBackoffUntilValid(
-        initTime: Long,
-        maxTime: Long,
-        isValidResult: (T) -> Boolean,
-        action: suspend () -> T,
-    ): T {
-        var result = action()
-        var sum = 0L
-        var waiting = initTime
-        while (!isValidResult(result) && sum < maxTime) {
-            delay(waiting)
-            sum += waiting
-            waiting *= 2
-            result = action()
-        }
-        return result
-    }
-
-    override fun <T> blockEnsureMinExecutionTime(
-        delayTime: Int,
-        action: () -> T,
-    ): T {
-        val start = System.currentTimeMillis()
-        val result = action()
-        val end = System.currentTimeMillis()
-
-        val remainingDelay = delayTime + start - end
-
-        if (remainingDelay > 0) {
-            Thread.sleep(remainingDelay)
-        }
-        return result
-    }
-
-    override fun <T> blockExponentialBackoffUntilValid(
-        initTime: Long,
-        maxTime: Long,
-        isValidResult: (T) -> Boolean,
-        action: () -> T,
-    ): T {
-        var result = action()
-        var sum = 0L
-        var waiting = initTime
-        while (!isValidResult(result) && sum < maxTime) {
-            Thread.sleep(waiting)
-            sum += waiting
-            waiting *= 2
-            result = action()
-        }
-        return result
-    }
-}
